# api/openapi.yaml
openapi: 3.1.0
info:
  title: MOINA API
  version: 0.1.0
  summary: AI 소셜 지식 네트워크 REST·SSE·MCP API
  description: |
    서버에 실제 등록된 canonical route 계약입니다. 브라우저 session 또는 개인 API/MCP key로 인증합니다.
    Cookie 인증의 unsafe 요청은 `moina_csrf` cookie와 같은 `X-CSRF-Token` header가 필요합니다.
  license: {name: MIT, identifier: MIT}
  contact: {name: MOINA, url: 'https://github.com/hkjang/moina'}
servers:
  - {url: '/', description: 같은 origin}
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 로컬 로그인
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses: {'200': {$ref: '#/components/responses/CurrentUser'}, '401': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: 관리자가 허용한 로컬 가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterInput'}
      responses: {'201': {$ref: '#/components/responses/CurrentUser'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
...
  /api/v1/auth/oidc/login:
    get:
      tags: [Auth]
      summary: Authorization Code + PKCE 시작
      security: []
      parameters:
        - {name: returnTo, in: query, schema: {type: string, pattern: '^/'}}
      responses: {'302': {description: OIDC provider로 이동}, '404': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC callback
      security: []
      parameters:
        - {name: code, in: query, required: true, schema: {type: string}}
        - {name: state, in: query, required: true, schema: {type: string}}
      responses: {'302': {description: 원래 앱 경로로 이동}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 현재 사용자와 유효 권한
      responses: {'200': {$ref: '#/components/responses/CurrentUser'}, '401': {$ref: '#/components/responses/Error'}}
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: session 종료
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/feed:
    get:
      tags: [Feed]
      summary: For Me 또는 Following feed
      parameters:
        - {name: mode, in: query, schema: {type: string, enum: [for_me, following], default: for_me}}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Limit'}
      responses: {'200': {$ref: '#/components/responses/PostPage'}}
  /api/v1/flows/{mode}:
    get:
      tags: [Feed]
      summary: 브랜드 언어 feed alias
      parameters:
        - {name: mode, in: path, required: true, schema: {type: string, enum: [for_me, following]}}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Limit'}
      responses: {'200': {$ref: '#/components/responses/PostPage'}}
  /api/v1/posts:
    get:
      tags: [Posts]
      summary: Moin 목록
      parameters: [{$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/PostPage'}}
    post:
      tags: [Posts]
      summary: Moin 또는 Echo 작성
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/CreatePost'}
      responses: {'201': {$ref: '#/components/responses/Post'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/posts/{postID}:
    parameters: [{$ref: '#/components/parameters/PostID'}]
    get:
      tags: [Posts]
      summary: Moin 상세
      responses: {'200': {$ref: '#/components/responses/Post'}, '404': {$ref: '#/components/responses/Error'}}
    patch:
      tags: [Posts]
      summary: 내 Moin 수정
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/UpdatePost'}
      responses: {'200': {$ref: '#/components/responses/Post'}, '403': {$ref: '#/components/responses/Error'}}
    delete:
      tags: [Posts]
      summary: 내 Moin 삭제
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/posts/{postID}/replies:
    get:
      tags: [Posts]
      summary: Echo 목록
      parameters: [{$ref: '#/components/parameters/PostID'}, {$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/PostPage'}}
  /api/v1/posts/{postID}/reactions:
    parameters: [{$ref: '#/components/parameters/PostID'}]
    post:
      tags: [Posts]
      summary: Signal 반응 추가
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [type], properties: {type: {type: string, enum: [like, useful, insight, question, verify]}}}}}
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Posts]
      summary: 내 반응 제거
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/posts/{postID}/reactions/{reaction}:
    delete:
      tags: [Posts]
      summary: 지정 반응 제거
      parameters:
        - {$ref: '#/components/parameters/PostID'}
        - {name: reaction, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/CSRF'}
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/posts/{postID}/bookmark:
    parameters: [{$ref: '#/components/parameters/PostID'}]
    post:
      tags: [Posts]
      summary: Pocket 저장
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Posts]
      summary: Pocket 제거
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/posts/{postID}/remoin:
    post:
      tags: [Posts]
      summary: Remoin
      parameters: [{$ref: '#/components/parameters/PostID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'201': {$ref: '#/components/responses/Post'}}
    delete:
      tags: [Posts]
      summary: Remoin 취소
      parameters: [{$ref: '#/components/parameters/PostID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/moins:
    get:
      tags: [Posts]
      summary: Moin 목록 alias
      responses: {'200': {$ref: '#/components/responses/PostPage'}}
    post:
      tags: [Posts]
      summary: Moin 작성 alias
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/CreatePost'}
      responses: {'201': {$ref: '#/components/responses/Post'}}
  /api/v1/moins/{postID}:
    get:
      tags: [Posts]
      summary: Moin 상세 alias
      parameters: [{$ref: '#/components/parameters/PostID'}]
      responses: {'200': {$ref: '#/components/responses/Post'}}
  /api/v1/users/{username}:
    get:
      tags: [Social]
      summary: 사용자 프로필
      parameters: [{$ref: '#/components/parameters/Username'}]
      responses: {'200': {$ref: '#/components/responses/User'}, '404': {$ref: '#/components/responses/Error'}}
  /api/v1/profiles/{username}:
    get:
      tags: [Social]
      summary: 사용자 프로필 alias
      parameters: [{$ref: '#/components/parameters/Username'}]
      responses: {'200': {$ref: '#/components/responses/User'}}
  /api/v1/links/{userID}:
    post:
      tags: [Social]
      summary: 사용자 Link
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Social]
      summary: Link 해제
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/users/{userID}/block:
    post:
      tags: [Social]
      summary: 사용자 차단
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Social]
      summary: 차단 해제
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/users/{userID}/mute:
    post:
      tags: [Social]
      summary: 사용자 음소거
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Social]
      summary: 음소거 해제
      parameters: [{$ref: '#/components/parameters/UserID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/profile:
    get:
      tags: [Profile]
      summary: 내 프로필
      responses: {'200': {$ref: '#/components/responses/User'}}
    patch:
      tags: [Profile]
      summary: 내 프로필 수정
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses: {'200': {$ref: '#/components/responses/User'}}
  /api/v1/profile/preferences:
    get:
      tags: [Profile]
      summary: 개인화·접근성 설정
      responses: {'200': {$ref: '#/components/responses/Data'}}
    put:
      tags: [Profile]
      summary: 개인화·접근성 설정 저장
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses: {'200': {$ref: '#/components/responses/Data'}}
  /api/v1/profile/password:
    post:
      tags: [Profile]
      summary: 로컬 비밀번호 변경
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses: {'204': {$ref: '#/components/responses/NoContent'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/profile/keys:
    get:
      tags: [Profile]
      summary: 내 API·MCP 키 목록
      responses: {'200': {$ref: '#/components/responses/Data'}}
    post:
      tags: [Profile]
      summary: 키 생성(원문 1회 표시)
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/APIKey'}
      responses: {'201': {$ref: '#/components/responses/Data'}}
  /api/v1/profile/keys/{keyID}:
    parameters: [{$ref: '#/components/parameters/KeyID'}]
    patch:
      tags: [Profile]
      summary: 키 권한·만료 변경
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/APIKey'}
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Profile]
      summary: 키 폐기
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/profile/keys/{keyID}/rotate:
    post:
      tags: [Profile]
      summary: 개인 키 원자적 회전
      parameters: [{$ref: '#/components/parameters/KeyID'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
  /api/v1/topics:
    get:
      tags: [Discovery]
      summary: Topic·Pulse 목록
      parameters: [{$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
  /api/v1/topics/{slug}:
    get:
      tags: [Discovery]
      summary: Topic 상세
      parameters: [{$ref: '#/components/parameters/Slug'}]
      responses: {'200': {$ref: '#/components/responses/Data'}, '404': {$ref: '#/components/responses/Error'}}
  /api/v1/topics/{slug}/follow:
    post:
      tags: [Discovery]
      summary: Topic Link
      parameters: [{$ref: '#/components/parameters/Slug'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/OptionalObject'}
      responses: {'200': {$ref: '#/components/responses/Data'}}
    delete:
      tags: [Discovery]
      summary: Topic Link 해제
      parameters: [{$ref: '#/components/parameters/Slug'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
  /api/v1/search:
    get:
      tags: [Discovery]
      summary: Moin·사용자·Topic·Moim 통합 검색
      parameters:
        - {name: q, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 100}}
        - {name: type, in: query, schema: {type: string, enum: [all, posts, users, topics, moims]}}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Limit'}
      responses: {'200': {$ref: '#/components/responses/Data'}}
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: 내 알림
      parameters: [{$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/Data'}}
  /api/v1/notifications/read:
    post:
      tags: [Notifications]
      summary: 알림 읽음 처리
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      requestBody: {$ref: '#/components/requestBodies/Object'}
      responses: {'204': {$ref: '#/components/responses/NoContent'}}
...
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: moina_session}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: MOINA API/MCP key}
  parameters:
    CSRF:
      name: X-CSRF-Token
      in: header
      required: false
      description: Cookie 인증 unsafe 요청에서 필수. Bearer key에는 적용하지 않습니다.
      schema: {type: string}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}}
...
  requestBodies:
    Login:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [username, password]
            properties:
              username: {type: string, minLength: 1, maxLength: 254}
              password: {type: string, minLength: 1, maxLength: 72, writeOnly: true}
...
  responses:
...
    CurrentUser:
      description: current user
      content: {application/json: {schema: {type: object, properties: {data: {$ref: '#/components/schemas/AuthView'}}}}}
...
    NoContent: {description: 성공, 응답 본문 없음}
    Error:
      description: 오류
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
security:
  - cookieAuth: []
  - bearerAuth: []